Fix bugs for automate_onboard_project on AQuA
